+++ pyshacl/constraints/__init__.py
@@ -4,9 +4,10 @@ from .value_constraints import ClassConstraintComponent, DatatypeConstraintCompo
 from .cardinality_constraints import MinCountConstraintComponent, MaxCountConstraintComponent
 from .value_range_constraints import MinExclusiveConstraintComponent, MinInclusiveConstraintComponent, MaxExclusiveConstraintComponent, MaxInclusiveConstraintComponent
 from .string_based_constraints import MinLengthConstraintComponent, MaxLengthConstraintComponent, PatternConstraintComponent
+from .property_pair_constraints import EqualsConstraintComponent, DisjointConstraintComponent, LessThanConstraintComponent, LessThanOrEqualsConstraintComponent
 from .logical_constraints import NotConstraintComponent, AndConstraintComponent, OrConstraintComponent, XoneConstraintComponent
 from .shape_based_constraints import NodeShapeComponent, PropertyShapeComponent
-from .other_constraints import ClosedConstraintComponent, InConstraintComponent
+from .other_constraints import ClosedConstraintComponent, InConstraintComponent, HasValueConstraintComponent
 
 ALL_CONSTRAINT_COMPONENTS = [
     ClassConstraintComponent,
@@ -25,10 +26,16 @@ ALL_CONSTRAINT_COMPONENTS = [
     MinLengthConstraintComponent,
     MaxLengthConstraintComponent,
     PatternConstraintComponent,
+    EqualsConstraintComponent,
+    DisjointConstraintComponent,
+    LessThanConstraintComponent,
+    LessThanOrEqualsConstraintComponent,
     NodeShapeComponent,
     PropertyShapeComponent,
     ClosedConstraintComponent,
-    InConstraintComponent
+    HasValueConstraintComponent,
+    InConstraintComponent,
+
 ]
 
 CONSTRAINT_PARAMETERS_MAP = {p: c for c in ALL_CONSTRAINT_COMPONENTS